testValidation: Add @skipif decorator
Allows skipping a test if it cannot run with the current environment.
The tests is not run in the skip condition is True, but it must pass if
the skip condition is false. This can be used to skip tests that are not
compatible on python 3 yet.

Example usage:

    from testValidation import skipif

    @skipif(six.PY3, "Needs porting to python 3")
    def test_rusty_python2_code(self):
        ...

This is a subset of pytest.mark.skipif.

Change-Id: I71a2ae262482edc8acf303f18eac6f9035e710bc
Signed-off-by: Nir Soffer <nsoffer@redhat.com>