tests: Remove invalid pyc file tests
moduleloader test was creating empty (invalid) "a.pyc" file. Existance
of an invalid "a.pyc" file may casue the test to fail if "a.pyc" file is
newer then the "a.py". This depends on the resolution of the file system
timestamps and the timing.

Since import of pyc files is not the subject of this test, I removed the
pyc file from the test data.

Change-Id: Iffd6371a5f4ace637ea99d4a91550a277cbf3f72
Signed-off-by: Nir Soffer <nsoffer@redhat.com>