xleases: Add and remove sanlock resource
When adding a lease, write a sanlock resource. When removing a lease,
since we don't have yet a sanlock API for removing a resource, write a
dummy resource with empty lockspace and resource name.

Sanlock is not available yet on python 3, but this module and the tests
are compatible with python 3 using fakesanlock.

Change-Id: I02a0f538f217481a4e8585fd6c8045e20d9b4c69
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>