compat: Import sanlock from vdsm.compat
On python2 this import does not change the behavior in any way; if
sanlock is missing, we raise ImportError.

On python3 this import returns a MissingModule instance. Any operation
on this module will raise compat.Unsupported exception. However, tests
using fakesanlock can still test the modules and keep them python3
compatible. Previously we could not test modules importing sanlock on
python3 because sanlock is not available.

Change-Id: I8cce8d7efd4ee739c6b0fd13c693b0dc5f1a78b4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>