xleases: Robust sanlock resources management
Treat sanlock resources on storage as the one and only truth.  Before
creating or removing sanlock resources, mark the lease record as stale.
If sanlock operation succeeded, mark the lease record as used or free.

If vdsm is killed, the host loses power or sanlock operation fails,
storage can be in this state:

- No change on storage
- Stale lease record, sanlock resource may exists or not
- Used lease record, sanlock resource exists
- Free lease record, sanlock resource does not exist

The next time we try to create or remove the same lease_id we can repair
the stale lease record by reading the sanlock resource from storage.
Repairing stale record is not implemented yet.

Change-Id: I78d9588a84b620f95ce917d93376060a898c96c7
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>