tests: Fix incorrect test cleanup
If TestPgrep failed, it was leaving zombie processes. Because the test
did not wait the child processes, we would get a SIGCHLD singal when
they terminated, 3 seconds after the test. If the signal was received
when the test framework try to write to stderr, another unrelated test
would fail with interrupted system call.

We see EINTR failures from time to time both in travis and when running
tests locally.

Change-Id: I3838673e047e885cdf57b46dfa58fccae48980c3
Signed-off-by: Nir Soffer <nsoffer@redhat.com>