tests: Fix flaky test on travis
TestPgrep is failing from time to time on traivs, but not on jenkins.
Looking at the test, it seems that starting 3 child processes and
pgreping for them immediately is racy, since we don't have a guarantee
that a child process is running when execCmd returns.

Attempt to fix by adding fragile timeout, this should be good enough for
this test.

Change-Id: I9fd87425ce2f2c0055c7e79e0aace3411afca568
Signed-off-by: Nir Soffer <nsoffer@redhat.com>