Safe publication of the queue to avoid job loss
Since the queue is instance is modified by another thread, we are
subjected to job loss. After the Q is being swapped with a new Q, there
is a window where the T1 adding jobs to the Q may see the old Q,
particularry after we iterated all over it in #performPendingOperations.
In that case the job is lost forever.

The easy fix would be to make the Q volatile and force the JVM to order
the thread reads but we should consider also to move back to the
non-swapping field technic, prior to commit
d771b038a8356166114cc6eeed247927f771c7cd

Bug-url: https://bugzilla.redhat.com/show_bug.cgi?id=1401585
Change-Id: I58b0ddb6ec664aec879fbc2b4b3ad51ab0dd7757
Signed-off-by: Roy Golan <rgolan@redhat.com>