Remove the wrong handling of long running tasks
The queue traversal made assumptions about tasks that may remain in the
Q. After the change to queue swapping, those tasks will never be treated
any more and also a quick check shows that ChainedOperations are totally
unused. Therefor the handling for ChainedOperations is dropped and the
class itself is marked as @Deprecated to be later removed.

Change-Id: Ic0b1f8e514e7ab6f79591549e7661aa4790ce62d
Signed-off-by: Roy Golan <rgolan@redhat.com>