xleases: Make VolumeLeaseStatus more general
VolumeLeaseStatus was a simple enum, hiding useful information about a
sanlock lease, such as the host ids owning the lease, the lease mode
(shared or exclusive), and the lease version. The additional information
can be used for better error handling, easier debugging, or improved
logging.

Note that currently we use only exclusive leases, and sanlock python
binding does not report the lease mode, so the reported "shared" state
is always False.

Replace the enum with a map exposing the lease details, and rename to
LeaseStatus so we can use it also for external leases.

Change-Id: I1e46bdce209a4375d9b69a2da6603768ebb7600d
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>