virt: Prevent hook races after a successful migration
When a migration is successfully finished, it's handled in two places:

1. In the source thread, where setDownStatus is called.
2. In the life cycle event handling thread where post-migration hooks are
   called.

libvirt documentation doesn't specify when the life cycle event is emitted,
whether before or after migrateToURI3 finishes.  This creates a possible source
of races, when a VM may be destroyed before the hooks are run or finished.  The
risk may be higher when we introduce (in a followup patch) automatic destroying
of a VM on the source after a successful post-copy migration.

We could solve the problem by moving the hooks to
SourceThread._finishSuccessfully.  But then we could still miss other life
cycle events, such as a notification about switching to post-copy migration,
that must be processed in order to handle and finish the migration correctly.

So a safer and simpler solution for now is introducing a synchronizing object
ensuring that we don't finish the migration in the source thread until the
final life cycle event is received and processed.

Change-Id: I92207376c18184dc002ee246953a4ed3a87bce78
Bug-Url: https://bugzilla.redhat.com/1354343
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>