StorageDomain.getInfo - report vg metadata device for block sd
When a block storage domain is created, we disable the lvm metadata
on all the domain pvs except to the first one.

As we intend to add support for operations that modify the vg
stracture (like reducing devices) - we want to prevent those
operations from being made on the device containing the vg metadata.

This patch adds the reporting of that info, so that the engine can
leverage the info to block such operations from being performed.

Change-Id: I4a7763d2ab7d796be633ecd69f661cba96e29dde
Signed-off-by: Liron Aravot <laravot@redhat.com>