hostdev: optimize _process_productinfo
Processing of productinfo is currently one of the slower tasks.
Biggest reason for that is the number of XML lookups. We optimize the
function to do a single lookup for element's children and iterate over
these.

The result is approximately 1/3 of the original computing power
required; profiling done with 30127 devices to process:

before:
ncalls  tottime  percall  cumtime  percall filename:lineno(function)
30127    0.093    0.000    0.976    0.000 /usr/lib/python2.7/site-packages/vdsm/hostdev.py:311(_process_productinfo)
after:
ncalls  tottime  percall  cumtime  percall filename:lineno(function)
30127    0.054    0.000    0.356    0.000 /usr/lib/python2.7/site-packages/vdsm/hostdev.py:311(_process_productinfo)

Change-Id: I63582d9b30639a44ced94229ed9d235644fab4eb
Bug-Url: https://bugzilla.redhat.com/1398572
Signed-off-by: Martin Polednik <mpolednik@redhat.com>