localFsSD.py: Add lock class for V4 domains
Commit 8c5f53 added the key 4 to
StorageDomainManifest._domainLockTable, but neglected to do the same
for LocalFsStorageDomainManifest._domainLockTable, thus preventing
local storage domains from using the V4 format (and failing attempts to
create/upgrade a local DC to the 4.1 compatibility level).

This patch adds the missing entry to the dictionary in order to rectify
this problem.

Change-Id: I193d316dff42ee3e93b5f697a87a74e762340647
Bug-Url: https://bugzilla.redhat.com/1399493
Signed-off-by: Allon Mureinik <amureini@redhat.com>