virt: Don't crash on migration stats retrieval after a finished job
Nothing prevents asking SourceThread for migration stats any time.  Part of the
migration stats retrieval is asking libvirt for completed job stats.  This is a
racy situation after a migration completes and multiple invocations of that
call may result in an exception when the VM is already gone.

The libvirt call result is not critical and its unavailability shouldn't
prevent us from getting other stats.  So let's catch the exception and just log
it.

Change-Id: If2083f6dd594eaa20d1bed0fc9c7535cad71146b
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>