storage: Fix SDM indirection for StorageDomain.releaseHostId
StorageDomain.releaseHostId has an awfully named parameter 'unused'
which controls whether a hostId should be released when there are still
active resource leases.  During the SDM indirection work I mistook this
name as convention for a deprecated parameter and did not pass it along.
This led to the default 'False' being used which is not always what we
want.  This patch fixes that issue and should resolve issues with Hosted
Engine and storage maintenance.

Bug-Url: https://bugzilla.redhat.com/show_bug.cgi?id=1401540
Signed-off-by: Adam Litke <alitke@redhat.com>
Change-Id: I7167d37ad2c879e1ed64d12591aebc833bb5ce11