xleases: Support SPM verbs without pool uuid
Legacy SPM verbs required a pool uuid and used to access the single pool
object. Replace the HSM.pools dict used to keep the one and only pool
with a HSM._pool class attribute.

To avoid the ugliness and racy self._pool is None checks, the pool is
always set; when we are connected to a storage pool, it will be a
StoragePool instance, and when we are disconnected, a DisconnectedPool
instance.

A DisconnectedPool object is a dummy storage pool that raises
StoragePoolNotConnected to make sure code is not calling the storage
pool when not connected.

Remove HSM._get_hostid helper since now self._pool.id have the same
effect.

Change-Id: I4917cf414c999a47d22e432c6be1e8993f3cf11f
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>