tests: Fix EINTR errors in checker tests
This patch fixes this random failure seen both locally and in travis:

======================================================================
ERROR: test_stopping_repr (storage_check_test.TestDirectioChecker)
----------------------------------------------------------------------
Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/nose/case.py", line 133, in run
    self.runTest(result)
  File "/usr/lib/python2.7/site-packages/nose/case.py", line 151, in runTest
    test(result)
  File "/usr/lib64/python2.7/unittest/case.py", line 431, in __call__
    return self.run(*args, **kwds)
  File "/usr/lib64/python2.7/unittest/case.py", line 406, in run
    result.addSuccess(self)
  File "/usr/lib/python2.7/site-packages/nose/proxy.py", line 165, in addSuccess
    self.result.addSuccess(self.test)
  File "/vdsm/tests/testlib.py", line 362, in addSuccess
    self._writeResult(test, 'OK', TermColor.green, '.', True)
  File "/vdsm/tests/testlib.py", line 355, in _writeResult
    self.stream.writeln()
  File "/usr/lib64/python2.7/unittest/runner.py", line 25, in writeln
    self.write('\n') # text-mode streams translate to \r\n if needed
IOError: [Errno 4] Interrupted system call
-------------------- >> begin captured stdout << ---------------------
<DirectioChecker /path stopping at 0x7f9f6a46e450>

--------------------- >> end captured stdout << ----------------------
-------------------- >> begin captured logging << --------------------
2016-12-02 21:52:33,048 DEBUG (MainThread) [storage.check] Checker '/path' started (check:233)
2016-12-02 21:52:33,048 DEBUG (MainThread) [storage.check] START check '/path' (delay=0.00) (check:283)
2016-12-02 21:52:33,052 DEBUG (MainThread) [storage.check] Checker '/path' stopping (check:250)
2016-12-02 21:52:33,053 DEBUG (MainThread) [storage.asyncevent] Closing <EventLoop running=False closed=False at 0x140322659557456> (asyncevent:215)
--------------------- >> end captured logging << ---------------------

The issue was not waiting for a checker underlying process. If the
process finished when the test framework was trying to write to stderr,
the write was interrupted by the SIGCHLD signal.

One of the idle tests was starting a checker, repeating the running
checker tests. The test was simplified to not start the checker.

The running checker tests were fixed by running the event loop after
starting the checker. The checker complete method stop the event loop
after the checker reaped the child process.

In the stopping tests, the checker was switched to STOPPING state. In
this state, it does not invoke the complete callback. The tests were
fixed by waiting until the checker was stopped in a helper thread.

Change-Id: If6e15fda0676f7133549f889189e4f327344e0f4
Signed-off-by: Nir Soffer <nsoffer@redhat.com>