net: NM - Introduce a higher level abstraction layer
Introducing nm.networkmanager as another abstraction layer, hiding the
implementation details (currently through dbus).

In this patch, a device class is exposed, enabling the caller to fetch a
device connections, the active connection and cleanup the connection by
deleting all non-active connections.

The initialization of the NM package is also exposed: Opening a
connection to the system bus.

Change-Id: I0ea6e650c237a65ac70592025c5f3e8b6cd55590
Signed-off-by: Edward Haas <edwardh@redhat.com>