net: NM - Introduce ifcfg2connection mapper
Given an ifcfg file, return the corresponding NM connection info.
The connection info includes the UUID and the connection path.

Note: The underlying dbus interface used is not documented. However, it
is used in the ifcfg initscripts, therefore we expect it to be supported
in the future.
The integration tests give an initial protection in case this API is
changed.

Change-Id: I917cb14ff3d2bdf7a1ba7d936310fddc56db1e4c
Signed-off-by: Edward Haas <edwardh@redhat.com>