tests: Terminate running jobs on exit
When the test runner terminated, we terminated the test watcher, but
forget to terminate the watcher sleep child process. This caused a delay
of more than 9 minutes after the tests completed when running in mock.

Now we register exit trap handler to terminate all jobs in both the main
test process and the watcher sub process.

Registering trap handler in a subshell does not work on EL7, so the
watcher code was extracted into a py-watch program.

With this change, vdsm check-patch builds takes 9 minutes instead of 19.

Change-Id: I28a1188ad3ddc6af76263281b63b6574c8f4f2fe
Signed-off-by: Nir Soffer <nsoffer@redhat.com>