vm: migration: handle destroy on startup on dest
When a pre-copy migration is not converging, the
source Vdsm may decide to abort the operation, and
during this process it destroys the destination side Vm.

This is part of the normal flow of operations, and one
of the expected and supported outcomes of the pre-copy
migrations.

What should not happen, however, is that the destination
side VM reports error in the logs. This is wrong for users,
which see a scary and pointless stacktrace, and may
confuse Engine if it sees a VM down in Error.

With this patch, we add explicit check for this condition,
and we let the destination VM to be down without error.

See
http://lists.ovirt.org/pipermail/users/2016-December/044477.html
For a recent case.

Change-Id: I9fa5d99ca10877bb2c230f892562c5ebee455e21
Bug-Url: https://bugzilla.redhat.com/1134974
Signed-off-by: Francesco Romani <fromani@redhat.com>