Cold Merge: Implement StoragePool.prepareMerge
This operation is required before performing (cold) merge.  Prepare
merge will calculate the required allocation for base volume, extend the
base volume or enlarge it (if the size of volume being removed is larger
than the base size), and mark it as ILLEGAL.

Note that testing extending of raw volumes is broken now, since we
cannot create a storage domain object in the tests. We will have to fix
this later, and currently this mus the verified manually.

Change-Id: I07614bd54c7da6f0a667cfc1ecfb0573588978e2
Bug-Url: https://bugzilla.redhat.com/1246114
Signed-off-by: Adam Litke <alitke@redhat.com>
Signed-off-by: Nir Soffer <nsoffer@redhat.com>
Signed-off-by: Ala Hino <ahino@redhat.com>