virt: Make sure to report 100% progress on completed migrations
We set migration progress to 100 in SourceThread when the migration is
successfully completed.  However there are timing issues and we can report less
than 100 when we report downtime from onJobCompleted.  Moreover the source
thread is not running if Vdsm got restarted during the migration and then 0
progress is reported.

This patch ensures that we always report 100% progress when downtime is
reported.

Change-Id: Iacf75ad547d2c97298d81a137216d23ab2c1f211
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>