Live Merge: Fix teardown volume support
When patches 7365ede8 (Live Merge: Teardown volume on HSM after live merge) and
68909ef3 (Live Merge: Remove volume run link after live merge) backportetd from
master to 4.0 branch, we didn't change the code to produce domain instead of
domain_manifest that isn't supported in 4.0. In this patch, code changed to
produce a domain and relevant methods (teardownVolume and removeVolumeRunLink)
on manifest class moved to domain class.

This patch is applied directly on 4.0 branch because obviously it is not
related to master.

Change-Id: I484aef0a5b392c273f7c9fb2b55ed4c708beeb2a
Bug-Url: https://bugzilla.redhat.com/1400137
Bug-Url: https://bugzilla.redhat.com/1399688
Signed-off-by: Ala Hino <ahino@redhat.com>