xleases: Cleanup free records writing and lookup
Previously free records contained the offset of the lease which is not
needed when there is no lease, and creation of free records was
duplicated in 3 places.

Change the free record format to have also an invalid offset (0), and
use a special EMPTY_RECORD constant for writing and searching free
records.

Change-Id: Id754f79cb3047bba442202e3f0d0a961f6b36103
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>