check-merged: global variable left unchanged due to tee pipe
components of a pipeline are executed in a subshell, causing any global
variable assignments to remain unchanged in the main shell.

as such, modifications to the global variable "failed" (which dictates
check-merged's return value) in the main shell were failing to occur.

instead of relying on a global variable, run_func_tests returns its own
return code which is later used to determine the return code of
check-merged.

additionally, the two test batches were separated into two lago shells;
running in a single shell caused the latter batch to override the result
of the former, consequently causing the return value of the entire
testing process to be solely determined by it.

Change-Id: If74f7a6d696acf2404afecf6fd1f3d2e28ab72cf
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>