xleases: Implement basic leases APIs
Implement the basic APIs required by engine to implement VM leases:
- Lease.create
- Lease.delete
- Lease.info

Lease.status and Lease.rebuild will be implemented later, as engine does
not need them right now. They may be useful for debugging and repairing
the external leases volume.

Creating and deleting a lease is implemented in the StorageDomain class,
since these operations must be called on the SPM. Getting lease info is
implemented on the StorageDomainManifest class so we don't have to
produce a (risky) StorageDomain on any host.

Creating and deleting a lease does not fail if the lease exists or was
already removed, since engine does not check SPM tasks errors. This can
be improved later when we switch this verbs to storage jobs.

Getting lease info require reading the external leases index on any
call, in case the SPM added or removed leases. On the SPM we can read
the index once and only write changes to storage, but this optimization
is not implemented yet, and I'm not sure that it is needed.

Change-Id: I9d8b38a7f940ebeee4a0f414c4c9474f9c6f1c5b
Bug-Url: https://bugzilla.redhat.com/1317429
Signed-off-by: Nir Soffer <nsoffer@redhat.com>