vmxml: test surprising behaviour of find_first
During the resolution of rhbz#1400137 , we stumbled on
a possibly surprising behaviour of vmxml.find_first:
We needed to traverse a nested structure like

  <disk>
    <source ... >...</source>
    <backingStore>
      <source ... >...</source>
      </backingStore/>
    </backingStore>
  </disk>

If we use vmxml.find_first, all worked fine until we run it
again in the innermost element, like

  new_elem = vmxml.find_first(elem, 'backingStore')

in this case we always had as result the innermost element itself,
instead of None or one exception. This lead to an infinite loop.

In this patch we add a @brokentest to document this behaviour.

Change-Id: I824f8f45fb95a36edff876f5bde474970d1f0341
Related-To: https://bugzilla.redhat.com/1400707
Signed-off-by: Francesco Romani <fromani@redhat.com>