vm: set _destroy_requested event onQemuDeath
We set the _destroy_requested event on externally-triggered
destroy requests, like VM.destroy() calls.

Actually, also the onQemuDeath is an externally-triggered
destruction request: it just comes from the system, not from the
Engine.

Uniform handling of the destruction requests allow us
to better react to complex events, like it is done in the
followup patch.

Change-Id: I9e9301849c6fcf56f8daf062706ce985871e5ded
Related-To: https://bugzilla.redhat.com/1134974
Signed-off-by: Francesco Romani <fromani@redhat.com>