sd: Move teardownVolume to StorageDomain
This method was added in master to StorageDomainManifest, used when
cleaning up after live merge. In 4.0 we still use StorageDomain for
cleanup, so the method must be implemented there.

This completes commit 198e48de58ac (Live Merge: Fix teardown volume
support) which did the same move for BlockStorageDomain, but missed the
default empty implementation in StorageDomain, used by file storage.

Change-Id: Idce11c424c03c1b04f1ed3c84baca85a675c80ea
Bug-Url: https://bugzilla.redhat.com/1400137
Bug-Url: https://bugzilla.redhat.com/1399688
Signed-off-by: Nir Soffer <nsoffer@redhat.com>