storage: add missing discard parameter
In commit 67bac28162c6 (storage: add discard argument to all related
flows) I added the discard parameter to fileVolume.delete and to
blockVolume.delete, but forgot to update one call to these methods,
which is a part of the rollback flow when creating a volume.

This patch fixes this method call by adding the new argument to it.

Change-Id: I506dbfb30b12c9167c0bd1b733a448baceb57f83
Signed-off-by: Idan Shaby <ishaby@redhat.com>