net: Optimize DHCP information reporting
A performance degradation has been introduced by DHCP state checks in
4.0.
This patch attempts to resolve the performance issue by collecting the
DHCP information in bulks, collecting it on a collection of devices in
one pass.

The DHCP information is collected for all detected devices on the host
and from the networks ifaces.

The networks ifaces are actually pointing to one of the system devices,
therefore, in a following path, the second dhcp info collection can be
removed and just copied from the correct device.

Change-Id: Ic4bf8c0763d8f525c0b94f5beb4fa2d56e804a04
Signed-off-by: Edward Haas <edwardh@redhat.com>