hsm: Fix SDM.set_volume_generation
When SDM.set_volume_generation was written the way to get the pool id in
sdm verbs was self._get_hostid(), but recently this helper was removed
and now the pool id is access using self._pool.id. Fixing pool id
revealed that the new api module was not imported.

Change-Id: Ia6ef7fba3b7683b6d22d7b9ae6e3a7054fdc580b
Signed-off-by: Nir Soffer <nsoffer@redhat.com>