volume: Remove LeaseStatus shared flag
Sanlock supports shared leases, but the python binding does not expose
the shared status of a lease, so we better not expose this value. Engine
is using only the owners field currently, so this change is safe.

This should have been part of https://gerrit.ovirt.org/67610 but it was
lost by mistake.

Change-Id: I7d8e112bbbe14861308042251a32793fb9de7090
Signed-off-by: Nir Soffer <nsoffer@redhat.com>