virt: Configurable name for guest agent channel
The channel name com.redhat.rhevm.vdsm for the communication
with the guest agent has been deprecated in favor of naming it
aligned with the upstream project name.

Since there is still the possibility that the legacy channel
name is still required, it will be now the responsibility of
the VM creator (caller of the VM.create API call) to specify
the name of the channel.

Bug-Url: https://bugzilla.redhat.com/1405378
Change-Id: Ia9f7dc2200ae3b022031f3a290983b6a4bcee952
Signed-off-by: Vinzenz Feenstra <vfeenstr@redhat.com>