spec: rpmlint: clean dir-or-file-in-var-run
Remove /var/run/* files from spec, as they do not survive reboot. One of
the files was missing from systemd's tmpfiles, so it was added there, to
be recreated on each reboot.

The patch runs rpmlint on any change of the spec file to make sure that
this rpmlint error would not return.

Change-Id: I0bdc3b1f2be960d536856465979c89ac85bce4a5
Signed-off-by: Dan Kenigsberg <danken@redhat.com>