storage: improve worker logs
In this patch, the worker logs are improved in order to ease debugging.
The correlation between the task ID and the thread running it, is now easy
to find.

When a task is started an info level log is printed with this format:
"<date> INFO (<thread-name>) START task <task-id> (cmd=<cmd>, args=<args>)"

When the task finishes an info level log is printed with this format:
"<date> INFO (<thread-name>) FINISH task <task-id>"

For example:
2016-12-19 17:38:13,149 INFO  (tasks/0)
[storage.ThreadPool.WorkerThread] START task
4c85331d-b424-40e5-80d8-ae0f57d092ed (cmd=<bound method Task.commit of
<storage.task.Task instance at 0x283d440>>, args=None) (threadPool:208)

2016-12-19 17:38:37,228 INFO  (tasks/0)
[storage.ThreadPool.WorkerThread] FINISH task
4c85331d-b424-40e5-80d8-ae0f57d092ed (threadPool:210)

Change-Id: I3ad2ff7cdc154a55856b3f57b0a41dcfb41d34f7
Signed-off-by: Fred Rolland <frolland@redhat.com>