clusterlock: Handle invalid lease name
Old cold merge flow used to create a temporary volume for the merge, and
once the merge is done, rename the temporary volume to the final volume
name. Unfortunately, this rename did not include renaming of the volume
lease, leaving invalid lease name on storage.

This issue is probably old as vdsm, but because we never used or queried
volume leases, the issue was hidden. In 4.1 we use volume leases in many
storage flows, and we report volume lease status in getVolumeInfo. When
querying lease status on a cold merge volume after the merge, sanlock is
raising an exception, failing the cold merge flow.

    SanlockException: (-227, 'Unable to read resource owners')

We plan to fix the old code merge flow, but the bad leases are out in
the wild, so we must handle them in a robust way when querying volume
lease status.

This patch adds the cluserlock.InvalidLeaseName exception for reporting
this error.  In getVolumeInfo, we log this exception and report that
this volume does not support lease.

Change-Id: Ibe0f1e15791ba9c2da84ab7f6ec2d9b2c1a42c9c
Bug-Url: https://bugzilla.redhat.com/1405817
Signed-off-by: Nir Soffer <nsoffer@redhat.com>