Frame split when using multiple hosts
When we use fake vdsm which is located on different host the os could
split network frames and we need to be prepared for it.

We had an issue that fakevdsm close connection due to incorrect
heartbeat default.

Fakevdsm processed heartbeats every reactor loop.

When fakevdsm socked failed to bind we did not see it.


Change-Id: If74e45113f94107d50dd34c428711efe46033182
Signed-off-by: Piotr Kliczewski <piotr.kliczewski@gmail.com>