net: Networking restoration done in a greedy mode
The networking restoration has been done in an 'all or nothing' mode:
In case a network/bond setup failed, it failed all networks and did not
continue trying other networks.
The end result: VDSM failed to come up on boot.

With this patch, the restoration process is done in a greedy mode,
trying setting each network/bond and continue even if one fails.

Note that with this new restoration mode, VDSM may come up with partial
networks configured (even with none).

Change-Id: I264b43195c0db039b911e658c71290372882d0fc
Bug-Url: https://bugzilla.redhat.com/1242532
Signed-off-by: Edward Haas <edwardh@redhat.com>