build: Require python-cpopen for build
We recently removed cpopen from the build requirements ([1]), since it
was used only during the tests, and we do not run the tests during the
build.

However after this we merged a patch creating html version of the
schema. This code does not use cpopen, but depends on the vdsm.compat
module which requires cpopen on Python 2.

This patch fixes the following exception when running:

   /usr/bin/python schema_to_html.py vdsm-api.yml vdsm-api.html

Traceback (most recent call last):
   File "schema_to_html.py", line 29, in <module>
     from api import vdsmapi
   File "/builddir/build/BUILD/vdsm-4.19.1/lib/api/vdsmapi.py", line 28, in <module>
     from vdsm import utils
   File "/builddir/build/BUILD/vdsm-4.19.1/lib/vdsm/utils.py", line 33, in <module>
     from .compat import pickle
   File "/builddir/build/BUILD/vdsm-4.19.1/lib/vdsm/compat.py", line 44, in <module>
     from cpopen import CPopen
ImportError: No module named cpopen

[1] https://gerrit.ovirt.org/#/c/65500

Change-Id: Iac8ce7f91eb09b210ef7fb4ade9b224c59d9b9af
Signed-off-by: Yaniv Bronhaim <ybronhei@redhat.com>