mailbox: Allow shorter wait timeout for testing
The storage thread pool uses a crappy method for checking if threads
should exit; each thread is waiting on the message queue op to
waitTimeout seconds and check if the thread was terminated.

For mailbox testing, we are using 0.1 seconds monitorInterval, to avoid
unneeded delays during the test, but when we stop the test, we wait
about 6 seconds until the mailbox thread pools are stopped. This is
cause by using constant 3 seconds wait time in the thread pool.

Compute the wait timeout using the monitor interval, so we keep the
runtime behavior the same, but can use shorter timeouts for testing,
speeding up the mailbox tests:

$ ./run_tests_local.sh storage_mailbox_test.py
tests.storage_mailbox_test.SPM_MailMonitorTests
    testThreadLeak                                              OK
tests.storage_mailbox_test.TestMailbox
    test_send_receive                                           OK

----------------------------------------------------------------------
Ran 2 tests in 0.608s

Before this change, the tests needed more then 6 seconds, and were
marked as slowtest.

Change-Id: Ifcfb94b5bae7f5e7c9d077c44b63594d10f79bc9
Signed-off-by: Nir Soffer <nsoffer@redhat.com>