netlink: monitor: join scanner thread on __exit__
In case a timeout occurs, the _scan() function ultimately ends, but
nothing explicitly waits on its control thread. It is harmless in the
case of netlink.monitor, but it is impolite to leak threads, even for a
short while.

This patch explicitly waits for the scan thread to finish before exiting
from an event monitor context manager.

The issue was spotted by running the tests with --with-thread-leak-check.

Change-Id: I9dd0097ad8d6ba5288e27b8f02bc62fc562030e9
Signed-off-by: Dan Kenigsberg <danken@redhat.com>
Signed-off-by: Petr Horáček <phoracek@redhat.com>