storage_mailbox_test: test send/receive
The added test sends an xtnd message from to an hsm mailbox, and waits
until the spm mailbox processes it (or timeout expires). It then asserts
that the expected message arrived, and only it.

In order to add the test, this patch sets the proper size and content to
the empty inbox and outbox. It also avoided randomness, to have an
expectable message.

Change-Id: I46be970d1cf0c9e495e69185d8eeb8ebb68d991b
Signed-off-by: Dan Kenigsberg <danken@redhat.com>