vdsm-tool: Add format-xleases command
This command can be use to format the xleases special volume if the
volume is corrupted or there is a need to add the volume on a
development setup.

The caller must know the path to the volume and make sure the volume is
accessible (activate the lv if needed).

Example usage:

    $ vdsm-tool format-xleases -h
    usage: vdsm-tool [-h] sd_id path

    positional arguments:
      sd_id       storage domain UUID
      path        path to xleases volume

    optional arguments:
      -h, --help  show this help message and exit

Creating xleases volume on file storage:

    PATH=/rhev/data-center/mnt/server:_export/sd-id/dom_md/xleases
    truncate -s 1G $PATH
    vdsm-tool format-xleases sd-id $PATH

Creating the xleases volume on block storage:

    lvcreate --name xleases --size 1G sd-id
    vdsm-tool format-xleases sd-id /dev/sd-id/xleases
    lvchange -an sd-id/xleases

Change-Id: I5d1a1a34eea80ddca01967c0a2da2e9e50e50656
Bug-Url: https://bugzilla.redhat.com/1656815
Signed-off-by: Nir Soffer <nsoffer@redhat.com>