hsm: adding pe_alloc_count/pe_count to getVgInfo
This patch adds to pe_alloc_count/pe_count fields to getVgInfo.
The engine accepts a list of devices to reduce from the domain from the user.
In order to validate that there's enough space to reduce all those
devices (all or nothing), we need to determine how much free PEs there
are in all the other devices of the vg.
Adding pe_alloc_count/pe_count to each reported  pv will provide the engine the
ability to determine how much space there is and to either allow or
block the operation.

Change-Id: I460616c88087d3d410511b45fd319765dacca416
Signed-off-by: Liron Aravot <laravot@redhat.com>