xleases: Implement rebuild_index
Rebuilding the index is very similar to formatting, but we read the
sanlock resources from disk and write records for the existing
resources.

When reading resources, we have to support two cases; a missing
resource, or a deleted resource (with empty resource name). Both mean
that there is no lease in this slot.

Change-Id: If32f53b9b427e521ec0b877233f395995cb8b2c4
Bug-Url: https://bugzilla.redhat.com/1654417
Signed-off-by: Nir Soffer <nsoffer@redhat.com>