xleases: Wire the rebuild_leases API
Lease.rebuild_leases starts a new SPM task, taking exclusive lock on the
external leases volume. The caller must wait until the task completes
and clear the task.

While the index is rebuilding, accessing it from other hosts will fail
with InvalidIndex error since the index is mark as "updating" before
rebuilding it.

If this operation fail, the index will be left in "updating" state and
the caller should try again to complete the rebuild.

Change-Id: Id250216e0ad4f9b54cd86ee328b5babc4d822f5c
Bug-Url: https://bugzilla.redhat.com/1656815
Signed-off-by: Nir Soffer <nsoffer@redhat.com>