net: introducing normalize_literal_addr
the method receives a valid ip address, and returns it in its literal
form.

ipv4 addresses are effectively remaining unchanged, however ipv6
addresses may need to be enclosed with square brackets (in accordance
with rfc2732) in order to able to be used in uri/urls.

an immediate use for this method is making sure qemu uri's will hold
valid addresses.


Change-Id: Ibb519853e1e0b7f3036845bada4db38053fafd35
Signed-off-by: Leon Goldberg <lgoldber@redhat.com>