net test: Fix test_ipv4_default_route default route assertion
On some functional test runs, the test fails on the default route
assertion, which checks that there is no default route at the end of
the test.

The checks is wrong, as it should actually check that the configured
default route is not the one on the setup.
The setup itself may have an original default route already or none.

Note: It seems that we do not (always) preserve the default route that
was originally set on the host. This should be fixed in future patches.

Change-Id: Iacfce818f8faa35d533bdd8c0d6a784b4fc026d1
Signed-off-by: Edward Haas <edwardh@redhat.com>