net: For route deletion, specify the relevant family
Deleting a route without specifing the family may result in a deletion
of both IPv4 and IPv6 routes.
Such a scenario explicitly occurs when deleting the 'default' route.

This patch allows the caller to specify an explicit family for the route
deletion.
iproute2 module is the only caller at the moment.

Change-Id: Id241c7938ad7fa9266f8d4761eae267e30f07ef4
Signed-off-by: Edward Haas <edwardh@redhat.com>