testValidation: Add @xfail decorator
This decorator should be used to mark good tests as expected failure. In
this case the test is good, but the code is broken, and cannot be fix
yet.

The test will skip with the reason message if the test fail, and fail if
the test succeeds, since this means the code is working and we can
remove this decorator.

This is a poor man implementation of pytest.mark.xfail, see
http://doc.pytest.org/en/latest/skipping.html

Usage::

    @xfail("why this test canonot pass now...")
    def test_broken_code(self):
	...

Change-Id: Ib1ff3996ced0362ca5b2f4718b46f0bac3e32fc8
Signed-off-by: Nir Soffer <nsoffer@redhat.com>