exception: Add ContextException
VdsmExcpetion subclasses generate unhelpful error messages with no
context about the error, making debugging harder on both vdsm and its
clients.

GeneralException attempts to improve this by adding *value argument, and
rendering the value with the message, but this create unclear error
messages:

    Storage pool not connected: ("d607ab66-6733-4ab4-9dcd-32e7fe61fa87",)

Some exceptions are overriding __init__ to add specific arguments and
formatting; this creates better messages, but it does not scale.

This patch adds a new ContextException with a reason and **kwargs
arguments.  The arguments are stored into a context dict, and formatted
with the static message.

Example usage:

    raise HotplugLeaseFailed(reason="Not hot enough", temperature=42)

Would render this message:

    Failed to hotplug lease: {"reason": "Not hot enough", "temperature": 42}

The real error should include more useful information, like the lease
details and the error message from libvirt.

When all exceptions are converted to use ContextException, we can merge
it back into VdsmException.

Change-Id: Id9b2bd8630ce5713123aa84042ea2110fd1dea1d
https://bugzilla.redhat.com/1408825
Signed-off-by: Nir Soffer <nsoffer@redhat.com>