lease: Implement lease.Device.is_attached_to
When hot-unplugging a lease, we need to wait until the device is
removed, so we must implement is_attached_to method.

Implemented using the lease id and the domain id, which define the lease
identity. Not clear if we should check also the lease path and offset,
and how (impossible) mismatch should be handled.

Change-Id: I733e33dd98b6e847b306186591fefd4a7ddbfbf2
Bug-Url: https://bugzilla.redhat.com/1408825
Signed-off-by: Nir Soffer <nsoffer@redhat.com>