vm: Add or remove a lease when a VM is running
Add the capability to plug or unplug a lease while a VM is running. This
will be used to move the lease to another domain if the current domain
need to be put in maintenance.

I stole the vm property from https://gerrit.ovirt.org/61475 since it is
too late to backport it to 4.1, and I fell too dirty to copy and paste
the vm fetching code in the new verbs.

Change-Id: Id791c848ab5813b19da0ae45172af89497ba580b
Bug-Url: https://bugzilla.redhat.com/1408825
Signed-off-by: Nir Soffer <nsoffer@redhat.com>