net: extract the configuration action from sourceroute
sourceroute and the configurators expose a dependency loop which is best
to drop.
In this patch, the sourceroute configure action is re-organized, such
that sourceroute will no longer be aware of the configurator. It will
just return the values that need to be used for the configuration and
the caller will forward them to the relevant configurator.

In the following patch, the remove action will also take a similar
approach and the module level dependency will be completely removed.

Change-Id: I3a6f700f16f103f14b1a50e436843adb8df57ecb
Signed-off-by: Edward Haas <edwardh@redhat.com>