net: remove sourceroute dependency on the configurators
Extracting the remove action from sourceroute as done with the config
action.

As a result, sourceroute no longer is dependent on the configurators and
no longer needs to call one.
A user fetches the current sourceroute configuration from the module and
uses it to call the relevant configurator remove method.

Change-Id: Ic3bd18eaab5afc43cebde5cb6b43d106ba6a2ccf
Signed-off-by: Edward Haas <edwardh@redhat.com>