rpmlint: ignore only limited set of errors
Currently, our rpms fail in 5 tests:

    explicit-lib-dependency
    no-binary
    zero-length
    non-readable
    non-standard-dir-perm

We do not have the time to fix these failures right now, but we want to
make sure that no new failure is added in the mean while.

This patch makes rpmlint ignore the currently-broken tests, and fail
loudly if a new one pops in.

Reviewed-on: https://gerrit.ovirt.org/#/c/69444
Change-Id: I1a6e21b61b9b31a07d7cd3af2ed9ab8506c46c5a
Signed-off-by: Dan Kenigsberg <danken@redhat.com>