pylint: Check changed modules in check-patch
Add pylint-diff target checking the modules modified by this patch.
Since this check is fast enough, we can run it for every patch in
check-patch.sh.

Since pylint will fail now for many modules, we do not fail the build
yet. When the current code is pylint-clean we can start failing the
build if a new error is detected.

Timing pylint and pylint-diff for typical patch:

    $ time make pylint
    ...
    real 1m31.133s
    user 1m29.934s
    sys 0m0.849s

    $ time make pylint-diff
    ...
    real 0m12.364s
    user 0m11.879s
    sys 0m0.262s

Change-Id: Ia458909a300c8555a84f3d7f1f8a410c0856e444
Signed-off-by: Nir Soffer <nsoffer@redhat.com>