net: Introduce ip.route module with an iproute2 driver
As part of the effort to partition the existing configurators into
focused topic drivers, ip.route is introduced.

It includes a basic abstration api with an iproute2 implemented using
the ipwrapper module.

It is aimed to be intially used by the sourceroute module as the default
way to configure a sourceroute.
(It cannot replace the ifcfg configurator, but has a potential to
replace the iproute2 configurator when configuring sourceroute, given we
introduce ip.rule module in the following patch)

Change-Id: I9aac45bdb2cb790e649a831b02a74f5c86f5717e
Signed-off-by: Edward Haas <edwardh@redhat.com>