net: Introduce ip.rule module with an iproute2 driver
As part of the effort to partition the existing configurators into
focused topic drivers, ip.rule is introduced.

It includes a basic abstration api with an iproute2 implemented
using the ipwrapper module.

It is aimed to be intially used by the sourceroute module as the
default way to configure a sourceroute.
(It cannot replace the ifcfg configurator, but can now replace the
iproute2 configurator when configuring sourceroute)

Change-Id: Ia414aa771bdda7205760c4ed3ab379c8286c7cb5
Signed-off-by: Edward Haas <edwardh@redhat.com>