gluster: Fix error when brick is on a btrfs subvolume
When a brick is on a btrfs subvolume, "gluster volume status detail"
doesn't return the device, blockSize, mntOptions or fsName.
cli._parseVolumeStatusDetail however expects them to be there so
we get a "KeyError: 'device'" every time details are asked for.

Those fields are now tested by calling get() instead of being
directly accessed.

Change-Id: I55c6a36aea46e6d7c2d85bff358e84c4e797d770
Bug-Url: https://bugzilla.redhat.com/1410283
Signed-off-by: George Joseph <gjoseph@digium.com>