tests: Fix flaky picklecopy benchmark
We have a benchmark for picklecopy, failing randomly on travis, when
picklecopy was less than twice faster compared with deepcopy. I don't
think it is useful to fail the test randomly because of this reason, and
I don't think picklecopy should be significantly faster then deepcopy to
be useful. This code is much simpler then deepcopy and even if it is not
significantly faster, I prefer to use picklecopy.

The tests prints now the results to make this easier to debug. To see
the timing run the tests with the -s option:

$ ./run_tests_local.sh utils_tests.py -s
...
utils_test.TestPickleCopy
    test_picklecopy_exact      OK
    test_picklecopy_faster     deepcopy: 0.194, picklecopy: 0.053 OK
...

Change-Id: Ice02137c9e6285401a9af29fd1b5ecd2a9c6a2cb
Signed-off-by: Nir Soffer <nsoffer@redhat.com>