virt: Update VM memory in Vm, not vmdevices
VM memory is a property of the VM, not of a particular device.  When memory
devices change, VM memory can change as well.  We currently update VM memory
from the device code.  This way VM memory can be updated automatically on
memory device updates, but then we handle a VM property from device code and
the Vm instance must be available to the updating code.

Let's keep the responsibilities separated and update VM memory only within Vm,
for the price of having to be careful not to forget about the call at all the
places where memory devices can change.

Change-Id: Ia6034453b1b65a22b348fbfbfaaa5b1f391fe1e9
Bug-Url: https://bugzilla.redhat.com/1228543
Signed-off-by: Milan Zamazal <mzamazal@redhat.com>