sp: Fix attaching of live storage domain
When attaching a live storage domain, like the hosted engine storage
domain, the host id is already acquired, and must not be released after
attaching the domain.

Releasing the host id will cause sanlock to expire the leases on this
domain and kill the hosted engine VM, holding a volume lease on this
domain.

We already had protection for live storage domain monitored by vdsm.
However after restarting vdsm (e.g. by host-deploy), there is a window
where vdsm does not know about the hosted engine storage domain, and if
engine was trying to auto-import the hosted engine storage domain at
this point vdsm was acquiring and releasing the host id, killing the
hosted engine VM.

Change-Id: Icf1165972fb0c328ef49852674e878617291fc36
Bug-Url: https://bugzilla.redhat.com/1409112
Signed-off-by: Nir Soffer <nsoffer@redhat.com>